\name{AlphaFoldQuery-package}
\alias{AlphaFoldQuery-package}
\alias{AlphaFoldQuery}
\docType{package}
\title{\packageTitle{AlphaFoldQuery}}
\description{\packageDescription{AlphaFoldQuery}}
\section{Package Content}{\packageIndices{AlphaFoldQuery}}
\author{\packageAuthor{AlphaFoldQuery}}
\section{Maintainer}{\packageMaintainer{AlphaFoldQuery}}
\keyword{package}
