% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PredictedAlignedError-class}
\alias{PredictedAlignedError-class}
\title{Predicted Aligned Error class}
\description{
Each structure prediction in AlphaFold is accompanied by a JSON file giving
the expected position error at each residue.  This class represents the 
contents of this file type.
}
\section{Slots}{

\describe{
\item{\code{residue1}}{The residue on which the protein structure is aligned.}

\item{\code{residue2}}{The residue the predicted error applies to.}

\item{\code{distance}}{The predicted aligned error for each residue pairing.}

\item{\code{max_predicted_error}}{The maximum possible value of the predicted aligned error.}
}}

