% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{readPAE}
\alias{readPAE}
\title{Read an AlphaFold predicted aligned error file}
\usage{
readPAE(input)
}
\arguments{
\item{input}{Character vector of length 1.  Can be either a complete URL
or local path to a predicted aligned error JSON file or a Uniprot protein 
ID.}
}
\value{
An object of class [PredictedAlignedError-class]
}
\description{
Read an AlphaFold predicted aligned error file
}
\examples{
## using a Uniprot ID
readPAE( "O82387" )

## using the full URL to a JSON file
readPAE( "https://alphafold.ebi.ac.uk/files/AF-O15552-F1-predicted_aligned_error_v1.json" )

}
